python3: fix host compilation on macOS
macOS needs to link to CoreFoundation for gettext to work. We reorder the
autoconf tests so CoreFoundation is added to LIBS earlier and the -lintl test
does not fail (which would exclude it from the final set of flags).

Btw. the whole test seems fishy: if compilation fails it does not mean -lintl is
not needed, for this we would need to test a gettext function without -lintl.
Basically two tests (with and without -lintl) are needed to properly diagnose
the situation. The test in question had been written in 2009 as a fix for the
https://bugs.python.org/issue6154 bug report.

This may be an issue only with non-Framework builds (since Homebrew manages to
build a framework-based Python on macOS without any special shenigans) but I had
not tested it.