admin/runit: add init.d scripts
The services are kept in `/tmp/service` by default so they have
to be recreated on each restart. To help with this the `init.d`
script runs all executable scripts in `/etc/service.d`. They may
create the services in any way they wish but a helper script
`make_runit_service` is provided to make it a bit easier.